feat: switch to locally maximal hyperbolic set (#2)
The previous implementation only made sense for maps which are globally hyperbolic. However, many important theorems require weaker assumptions and conclusions, that a map is hyperbolic only on some subset. We switch to a design allowing this. Instead of a typeclass, we work with a structure that can be instantiated with various subsets even for a single map.